% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_direction_to_degrees.R
\name{cardinal_direction_to_degrees}
\alias{cardinal_direction_to_degrees}
\title{Function to transform cardinal direction wind directions to decimal degrees.}
\usage{
cardinal_direction_to_degrees(x, north_is = 360)
}
\arguments{
\item{x}{Character vector with cardinal direction elements.}

\item{north_is}{What decimal degree should north be transformed to? Can be 
either 360 or 0.}
}
\value{
Numeric vector.
}
\description{
Function to transform cardinal direction wind directions to decimal degrees.
}
\examples{
 

# Function uses uppercase and abbreviated cardinal directions
x <- c(
  "N", "NNE", "NE", "ENE", "E", "ESE", "SE", "SSE", "S", "SSW", "SW", "WSW", 
  "W", "WNW", "NW", "NNW"
)

# Transform to decimal degrees
cardinal_direction_to_degrees(x)

# Transform but use 0 degrees for north
cardinal_direction_to_degrees(x, north_is = 0)

}
\author{
Stuart K. Grange.
}
