% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_europe_data.R
\name{get_europe_data}
\alias{get_europe_data}
\title{Function to get European data for WACL (Wolfson Atmospheric Laboratory) uses.}
\usage{
get_europe_data(site)
}
\arguments{
\item{site}{Site(s) to get data for. \code{site} is a site code such as 
\code{"ch0001a"}. \code{\link{get_europe_sites}} can be used to find the 
sites available.}
}
\value{
Data frame.
}
\description{
Data imported by this function are sourced from the European AirBase and 
e-Reporting data repositories. \code{get_europe_data} will attempt to 
find any data with the sites used, but this does not guarantee that these 
sites' data are available.
}
\details{
This function is very simple and does not allow for control of dates and
variables yet. If useful, the function will be developed further.
}
\examples{
\dontrun{

# Get some swiss data, two sites
data_ch <- get_europe_data(site = c("ch0001a", "ch0038a"))

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{get_europe_sites}}
}

