% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_europe_sites.R
\name{get_europe_sites}
\alias{get_europe_sites}
\title{Function to get European's sites for WACL (Wolfson Atmospheric Laboratory) 
uses.}
\usage{
get_europe_sites()
}
\value{
Data frame.
}
\description{
\code{\link{get_europe_sites}} is used to get the monitoring sites which are 
supported by \code{\link{get_europe_data}}.
}
\examples{
\dontrun{

# Get sites
data_sites <- get_europe_sites()

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{get_europe_data}}
}

