% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_invalidations.R
\name{get_wacl_invalidations}
\alias{get_wacl_invalidations}
\title{Function to get WACL's invalidations.}
\usage{
get_wacl_invalidations(json = FALSE)
}
\arguments{
\item{json}{Should the return be in JSON format? Default is \code{FALSE}.}
}
\value{
Data frame or pretty printed JSON with correct data types.
}
\description{
\code{\link{get_wacl_invalidations}} is used to get date ranges where data 
is considered invlid. The invalidation table and ranges is dynamic so may 
change at anytime.
}
\examples{
\dontrun{

# Get sites
data_invalidations <- get_wacl_invalidations()

# Or print as json
get_wacl_invalidations(json = TRUE)

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{get_wacl_data}}, \code{\link{get_wacl_sites}},
\code{\link{get_wacl_processes}}
}

