% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_processes.R
\name{get_wacl_processes}
\alias{get_wacl_processes}
\title{Function to get WACL's processes.}
\usage{
get_wacl_processes()
}
\value{
Data frame with correct data types.
}
\description{
Function to get WACL's processes.
}
\examples{
\dontrun{

# Get processes
data_processes <- get_wacl_processes()

}

}
\author{
Stuart K. Grange
}

