% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_sites.R
\name{get_wacl_sites}
\alias{get_wacl_sites}
\title{Function to get WACL's sites.}
\usage{
get_wacl_sites()
}
\value{
Data frame with correct data types.
}
\description{
Function to get WACL's sites.
}
\examples{
\dontrun{

# Get processes
data_sites <- get_wacl_sites()

}

}
\author{
Stuart K. Grange
}

