% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icad.R
\name{read_ICAD}
\alias{read_ICAD}
\title{read_ICAD}
\usage{
read_ICAD(ch1, ch2 = NULL, tz = "UTC", pad = FALSE)
}
\arguments{
\item{ch1}{path to a channel 1 file}

\item{ch2}{(optional) path to a channel 2 file. If not supplied the string "Channel1" will be replaced with "Channel2" from \code{ch1}}

\item{tz}{time zone to use when formatting date. default "UTC"}

\item{pad}{Should NO2, NO2x and NOmeasured channels be linealy interpolated between the different file times. Default false}
}
\description{
Simple function to read ICAD data.
}
\author{
W. S. Drysdale
}
