% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_syft_data.R
\name{read_syft_data}
\alias{read_syft_data}
\title{Function to read Syft \code{.csv} files.}
\usage{
read_syft_data(file, verbose = FALSE)
}
\arguments{
\item{file}{Vector of Syft \code{.csv} file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Named list containing data frames.
}
\description{
Function to read Syft \code{.csv} files.
}
\details{
The Syft instrument reports time with sub second accuracy. This is 
preserved but to view this extra precision, set the \code{digits.secs} option.
}
\examples{
\dontrun{

# Load a data file
file <- "data/Participant 3/Breath Romance-Participant 3-20180205-152112.csv"
list_syft <- read_syft_data(file)

# What units does the list contain? (purrr needs to be loaded)
map(list_syft, names) \%>\% 
  flatten_chr()
  
# Extract observations
data_observations <- map_dfr(
  list_syft, 
  `[[`, 
  "detailed_concentrations", 
  .id = "file"
)

}

}
\seealso{
\code{\link{set_sub_second_option}}
}
\author{
Stuart K. Grange
}
