% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sub_second_option.R
\name{set_sub_second_option}
\alias{set_sub_second_option}
\title{Function to set R's options to print fractional seconds.}
\usage{
set_sub_second_option(digits = 5)
}
\arguments{
\item{digits}{Number of fractional seconds to display.}
}
\value{
Invisible, a change in R's global options.
}
\description{
Function to set R's options to print fractional seconds.
}
\author{
Stuart K. Grange
}
