% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_faam_netcdf.R
\name{tidy_faam_netcdf}
\alias{tidy_faam_netcdf}
\title{Function to transform a \code{NetCDF} file from FAAM activites to a data 
frame.}
\usage{
tidy_faam_netcdf(file)
}
\arguments{
\item{file}{File names of FAAM \code{NetCDF} files.}
}
\value{
Data frame.
}
\description{
Function to transform a \code{NetCDF} file from FAAM activites to a data 
frame.
}
\examples{
\dontrun{

# Create tidy data from a single file
data_faam <- tidy_faam_netcdf(file = "data/core_faam_1hz.nc")

# Transform multiple files
files <- c("data/core_faam_1hz.nc", "data/core_faam_10hz.nc")
data_faam_many <- tidy_faam_netcdf(file = files)

}

}
\author{
Stuart K. Grange
}
