% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{time_zone}
\alias{time_zone}
\title{Function to extract time-zone from a POSIXct date.}
\usage{
time_zone(date)
}
\arguments{
\item{date}{Date vector}
}
\value{
Character vector with a length of one.
}
\description{
Function to extract time-zone from a POSIXct date.
}
\examples{

# A date vector
x <- lubridate::ymd_h("2018-05-05 11", tz = "Europe/Zurich")
time_zone(x)

}
\author{
Stuart K. Grange
}
