% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_date_span.R
\name{aggregate_by_date_span}
\alias{aggregate_by_date_span}
\title{Aggregate Met Data from date Span}
\usage{
aggregate_by_date_span(df, df_met, progress = "none")
}
\arguments{
\item{df}{dataframe containing span range}

\item{df_met}{dataframe containg met data}

\item{progress}{supply "time" for progress bar}
}
\value{
Data frame
}
\description{
Given averaged data with span ranges in \code{"date_start"} and 
\code{"date_end"} columns and met data with "date" column, average met data 
over span range and append to data frame
}
\examples{
\dontrun{

# Combine GC and met data
data_all <- aggregate_by_date_span(data_gc, data_met)

}
 
}
\author{
Will Drysdale
}
