% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_date_span.R
\name{aggregate_by_date_span}
\alias{aggregate_by_date_span}
\title{Aggregate meteorological data from date span.}
\usage{
aggregate_by_date_span(df, df_met, warn = TRUE, progress = NULL, ...)
}
\arguments{
\item{df}{Data frame containing span range.}

\item{df_met}{Data frame containg met data.}

\item{warn}{Should the function give warnings?}

\item{progress}{Depreciated as purrr::map_Df is used over plyr::adply}

\item{...}{Additional arguments to pass to openair::timeAverage}
}
\value{
Data frame.
}
\description{
Given averaged data with span ranges in \code{"date_start"} and 
\code{"date_end"} columns and met data with "date" column, average met data 
over span range and append to data frame
}
\examples{
\dontrun{

# Combine GC and met data
data_all <- aggregate_by_date_span(data_gc, data_met)

}
 
}
\author{
Will Drysdale and Stuart K. Grange
}
