% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_data.R
\name{get_wacl_data}
\alias{get_wacl_data}
\title{Function to get data for WACL (Wolfson Atmospheric Laboratory) uses.}
\usage{
get_wacl_data(site, year)
}
\arguments{
\item{site}{Site(s) to get data for. Currently \code{"kirb"} and \code{"litp"} 
are supported.}

\item{year}{Year(s) to get data for.}
}
\value{
Tidy data frame contains sites' monitoring data with correct data 
types.
}
\description{
The data imported by this function should always be regarded as preliminary 
and not validated or ratified. \code{get_wacl_data} will attempt to find any
data with site and year combinations which are used, but this does not 
guarantee that these data are available.
}
\examples{
\dontrun{

# Get data for the kirb site for 2016
data_kirb_2016 <- get_wacl_data(site = "kirb", year = 2016)

# Get data for two sites for many years
data_two_sites <- get_wacl_data(site = c("kirb", "litp"), year = 2015:2017)

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{get_wacl_sites()}}
}

