% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_data.R
\name{get_wacl_data}
\alias{get_wacl_data}
\title{Function to get data for WACL (Wolfson Atmospheric Laboratory) uses.}
\usage{
get_wacl_data(site, year = 2015:2018, period = "hour")
}
\arguments{
\item{site}{Site(s) to get data for. Use \code{\link{get_wacl_sites}} to
find sites which are available. \code{site} is a site code such as 
\code{"kirb"}.}

\item{year}{Year(s) to get data for. A vector of integers, default is 
\code{2015:2018}.}

\item{period}{Aggregation period to get data for. Default is \code{"hour"}.
\code{"source"} can be used to get "source" data. Beware that the importing 
of source data can be slow because it is common for sites to have several 
million observations for a year.}
}
\value{
Data frame containing sites' monitoring data with correct data types.
}
\description{
The data imported by this function should always be regarded as preliminary 
and are often not validated or ratified. \code{get_wacl_data} will attempt to 
find any data with the site and year combinations used, but this does not 
guarantee that these data are available.
}
\details{
This function will return different table (data frame) formats depending on 
the \code{period} argument.
}
\examples{
\dontrun{

# Get hourly data for a single site with default years
data_litp <- get_wacl_data(site = "litp")

# Get hourly data for the kirb site for 2016
data_kirb_2016 <- get_wacl_data(site = "kirb", year = 2016, period = "hour")

# Get hourly data for two sites for many years
data_two_sites <- get_wacl_data(site = c("kirb", "litp"), year = 2015:2017, 
                                period = "hour")
                                
# Get source data, one-minute data for a site
data_kirb_source <- get_wacl_data(site = "kirb", year = 2016, period = "source")

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{get_wacl_sites}}, \code{\link{get_wacl_processes}}
}

