% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_processes.R
\name{get_wacl_processes}
\alias{get_wacl_processes}
\title{Function to get WACL's processes.}
\usage{
get_wacl_processes(json = FALSE)
}
\arguments{
\item{json}{Should the return be in JSON format? Default is \code{FALSE}.}
}
\value{
Data frame or JSON with correct data types.
}
\description{
Function to get WACL's processes.
}
\examples{
\dontrun{

# Get processes
data_processes <- get_wacl_processes()

# Or print as json
get_wacl_processes(json = TRUE)

}

}
\author{
Stuart K. Grange
}

