% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wacl_sites.R
\name{get_wacl_sites}
\alias{get_wacl_sites}
\title{Function to get WACL's sites.}
\usage{
get_wacl_sites(json = FALSE)
}
\arguments{
\item{json}{Should the return be in JSON format? Default is \code{FALSE}.}
}
\value{
Data frame or pretty printed JSON with correct data types.
}
\description{
\code{\link{get_wacl_sites}} is used to get the monitoring sites which are 
supported by \code{\link{get_wacl_data}}.
}
\examples{
\dontrun{

# Get sites
data_sites <- get_wacl_sites()

# Or print as json
get_wacl_sites(json = TRUE)

}

}
\seealso{
\code{\link{get_wacl_data}}
}
\author{
Stuart K. Grange
}
